pyproject.toml
[tool.poetry]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = ["Nicholas <sousa.nicholas2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "*"
pytest = "*"
pytest-mock = "*"
flake8 = "^7.1.1"
psycopg2-binary = "*"
gunicorn = "*"



[tool.pytest.ini_options]
addopts = "-ra -vvv"
testpaths = ["tests"]
pythonpath = ["src"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
